pluck hole
we need to know which functions had been exported after initial compilation. we have to execute the code to find out so valid imports and function definitions are executed to gather the exports. then, validation method runs through the imports to check for invalid imports using the gathered exports as a filter